backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
redis>=5.0.0
celery>=5.4.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<5.0.0
python-jose[cryptography]>=3.3.0
pytest>=8.0.0
httpx>=0.27.0
langchain>=1.2.12
langchain-community>=0.4.1
langchain-core>=1.2.19
duckduckgo-search>=7.3.0

